It is now not possible to join public/private games that you shouldn't have access to
There was a problem when games were in progress or complete that they were still technically eligible to join. There was also an issue where you could join a private game without the name and password by calling the endpoint for joining a public game if you knew the id.

These situations now all throw the correct type of exception and disallow you from joining