All navigation between scenes now happens with URLs
Moved routing logic into a plugin that (with a little bit of work) can be extracted into a standalone plugin

Unfortunately I couldn't find a way of satisfying TypeScript that `this.router` existed in each scene (as provided by Phaser) so had to resort to using a BaseScene. I wish there was a better way to handle this here.